+++ src/derivatives_trading_usds_futures/rest_api/apis/market_data_api.rs
@@ -31,6 +31,10 @@ const HAS_TIME_UNIT: bool = false;
 
 #[async_trait]
 pub trait MarketDataApi: Send + Sync {
+    async fn adl_risk(
+        &self,
+        params: AdlRiskParams,
+    ) -> anyhow::Result<RestApiResponse<models::AdlRiskResponse>>;
     async fn basis(
         &self,
         params: BasisParams,
@@ -1055,6 +1059,29 @@ impl std::str::FromStr for TopTraderLongShortRatioPositionsPeriodEnum {
     }
 }
 
+/// Request parameters for the [`adl_risk`] operation.
+///
+/// This struct holds all of the inputs you can pass when calling
+/// [`adl_risk`](#method.adl_risk).
+#[derive(Clone, Debug, Builder, Default)]
+#[builder(pattern = "owned", build_fn(error = "ParamBuildError"))]
+pub struct AdlRiskParams {
+    ///
+    /// The `symbol` parameter.
+    ///
+    /// This field is **optional.
+    #[builder(setter(into), default)]
+    pub symbol: Option<String>,
+}
+
+impl AdlRiskParams {
+    /// Create a builder for [`adl_risk`].
+    ///
+    #[must_use]
+    pub fn builder() -> AdlRiskParamsBuilder {
+        AdlRiskParamsBuilder::default()
+    }
+}
 /// Request parameters for the [`basis`] operation.
 ///
 /// This struct holds all of the inputs you can pass when calling
@@ -2144,6 +2171,33 @@ impl TopTraderLongShortRatioPositionsParams {
 
 #[async_trait]
 impl MarketDataApi for MarketDataApiClient {
+    async fn adl_risk(
+        &self,
+        params: AdlRiskParams,
+    ) -> anyhow::Result<RestApiResponse<models::AdlRiskResponse>> {
+        let AdlRiskParams { symbol } = params;
+
+        let mut query_params = BTreeMap::new();
+
+        if let Some(rw) = symbol {
+            query_params.insert("symbol".to_string(), json!(rw));
+        }
+
+        send_request::<models::AdlRiskResponse>(
+            &self.configuration,
+            "/fapi/v1/symbolAdlRisk",
+            reqwest::Method::GET,
+            query_params,
+            if HAS_TIME_UNIT {
+                self.configuration.time_unit
+            } else {
+                None
+            },
+            false,
+        )
+        .await
+    }
+
     async fn basis(
         &self,
         params: BasisParams,
@@ -3244,6 +3298,33 @@ mod tests {
 
     #[async_trait]
     impl MarketDataApi for MockMarketDataApiClient {
+        async fn adl_risk(
+            &self,
+            _params: AdlRiskParams,
+        ) -> anyhow::Result<RestApiResponse<models::AdlRiskResponse>> {
+            if self.force_error {
+                return Err(
+                    ConnectorError::ConnectorClientError("ResponseError".to_string()).into(),
+                );
+            }
+
+            let resp_json: Value = serde_json::from_str(
+                r#"{"symbol":"BTCUSDT","adlRisk":"low","updateTime":1597370495002}"#,
+            )
+            .unwrap();
+            let dummy_response: models::AdlRiskResponse = serde_json::from_value(resp_json.clone())
+                .expect("should parse into models::AdlRiskResponse");
+
+            let dummy = DummyRestApiResponse {
+                inner: Box::new(move || Box::pin(async move { Ok(dummy_response) })),
+                status: 200,
+                headers: HashMap::new(),
+                rate_limits: None,
+            };
+
+            Ok(dummy.into())
+        }
+
         async fn basis(
             &self,
             _params: BasisParams,
@@ -4042,6 +4123,69 @@ mod tests {
 
             Ok(dummy.into())
         }
+    }
+
+    #[test]
+    fn adl_risk_required_params_success() {
+        TOKIO_SHARED_RT.block_on(async {
+            let client = MockMarketDataApiClient { force_error: false };
+
+            let params = AdlRiskParams::builder().build().unwrap();
+
+            let resp_json: Value = serde_json::from_str(
+                r#"{"symbol":"BTCUSDT","adlRisk":"low","updateTime":1597370495002}"#,
+            )
+            .unwrap();
+            let expected_response: models::AdlRiskResponse =
+                serde_json::from_value(resp_json.clone())
+                    .expect("should parse into models::AdlRiskResponse");
+
+            let resp = client.adl_risk(params).await.expect("Expected a response");
+            let data_future = resp.data();
+            let actual_response = data_future.await.unwrap();
+            assert_eq!(actual_response, expected_response);
+        });
+    }
+
+    #[test]
+    fn adl_risk_optional_params_success() {
+        TOKIO_SHARED_RT.block_on(async {
+            let client = MockMarketDataApiClient { force_error: false };
+
+            let params = AdlRiskParams::builder()
+                .symbol("symbol_example".to_string())
+                .build()
+                .unwrap();
+
+            let resp_json: Value = serde_json::from_str(
+                r#"{"symbol":"BTCUSDT","adlRisk":"low","updateTime":1597370495002}"#,
+            )
+            .unwrap();
+            let expected_response: models::AdlRiskResponse =
+                serde_json::from_value(resp_json.clone())
+                    .expect("should parse into models::AdlRiskResponse");
+
+            let resp = client.adl_risk(params).await.expect("Expected a response");
+            let data_future = resp.data();
+            let actual_response = data_future.await.unwrap();
+            assert_eq!(actual_response, expected_response);
+        });
+    }
+
+    #[test]
+    fn adl_risk_response_error() {
+        TOKIO_SHARED_RT.block_on(async {
+            let client = MockMarketDataApiClient { force_error: true };
+
+            let params = AdlRiskParams::builder().build().unwrap();
+
+            match client.adl_risk(params).await {
+                Ok(_) => panic!("Expected an error"),
+                Err(err) => {
+                    assert_eq!(err.to_string(), "Connector client error: ResponseError");
+                }
+            }
+        });
     }
 
     #[test]